refactor(agents): extract attempt tool-policy helpers (#72072)
(cherry picked from commit 435712a2f476647f70be4e6fa61b8c1802a72437)
